Rewrite DWARF support
This patch rewrites most of the DWARF implementation. It fixes major
shortcomings:

- No support for multi-compilation units (for example multiple C files).
- No support for namespaced functions (makes rust functions confusing).
- Missing name linkage mapping (which was pulled out of the wazero
  implementation).

This comes at a cost of a lot more memory allocation, at the time of building
the the profile and during initialization, so outside of the profiling hot path.
Once the featureset of the DWARF implementation stabilizes, there are many
opportunities to speeding this up.