sampler: do not share sampling state across function calls (#73)
This PR is another follow-up from #71, which contains two fixes:

The first fix consists in not applying the sampling scale to time
measures of the CPU profiler. This is needed because the time deltas
between call and return of functions are not affected by the sampling
rate, it's an absolute value.

The second fix changes the sampling mechanism to track state per
function instead of sharing it across all functions, which was causing
too much skew. We now record `1 / sample rate` calls to each function
instead of `1 / sample rate` function calls, which gives a much more
accurate profiles.

---------

Signed-off-by: Achille Roussel <achille.roussel@gmail.com>