dwarf: fix inline calls reporting
Finally implement the original TODO of checking whether the PC being symbolized
actually belong to the the inline function to add to locations, instead of
adding all functions inlined in the caller, regardless of location. Also,
correctly classify whether a location is the result of an inline call.

The rust test is updated for the same reason.

Also make sure tests are compiled with -g.